# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HttpAuthorizationFailedError:
      additionalProperties: false
      description: A custom exception body model.
      properties:
        data:
          $ref: '#/components/schemas/HttpAuthorizationFailedErrorData'
        description:
          description: A human readable message to the client explaining the cause
            of the exception.
          title: Description
          type: string
        exception_id:
          enum:
          - authorizationFailedError
          title: Exception Id
          type: string
      required:
      - data
      - description
      - exception_id
      title: HttpAuthorizationFailedError
      type: object
    HttpAuthorizationFailedErrorData:
      description: 'An empty model used as default for describing exception data.

        Please overwrite this to define your own data model.'
      properties: {}
      title: HttpAuthorizationFailedErrorData
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: A service exposing an external API to commission file deletionsfrom
    the wholefile backend.
  title: Purge Controller Service
  version: 0.1.0
openapi: 3.0.2
paths:
  /files/{file_id}:
    delete:
      description: Send out an event to delete the file with the given id.
      operationId: deleteFile
      parameters:
      - in: path
        name: file_id
        required: true
        schema:
          title: File Id
          type: string
      responses:
        '202':
          content:
            application/json:
              schema: {}
          description: Commissioned file deletion
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpAuthorizationFailedError'
          description: 'Bearer token could not be validated.

            Exceptions by ID:

            - authorizationFailedError: Bearer token could not be validated.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Deletes the corresponding file.
      tags:
      - PurgeControllerService
  /health:
    get:
      description: Used to test if this service is alive
      operationId: health_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health
      tags:
      - PurgeControllerService
servers:
- url: /
tags:
- name: PurgeControllerService